Move generate test keyring function to utils
