+++ test/packages/utils.js
@@ -16,6 +16,10 @@
     along with the caver-js. If not, see <http://www.gnu.org/licenses/>.
 */
 
+const Keyring = require('../../packages/caver-wallet/src/keyring/keyring')
+const { KEY_ROLE } = require('../../packages/caver-wallet/src/keyring/keyringHelper')
+const PrivateKey = require('../../packages/caver-wallet/src/keyring/privateKey')
+
 const unitMap = {
     peb: '1',
     kpeb: '1000',
@@ -30,6 +34,42 @@ const unitMap = {
     GKLAY: '1000000000000000000000000000',
 }
 
+const generateDecoupledKeyring = () => {
+    const keyring = Keyring.generate()
+    keyring.key = PrivateKey.generate()
+    return keyring
+}
+
+const generateMultiSigKeyring = (num = 3) => {
+    const keyring = Keyring.generate()
+    const multipleKeys = []
+    for (let i = 0; i < num; i++) {
+        multipleKeys.push(PrivateKey.generate())
+    }
+    keyring.key = multipleKeys
+    return keyring
+}
+
+const generateRoleBasedKeyring = numArr => {
+    if (numArr === undefined) {
+        numArr = Array(KEY_ROLE.ROLE_LAST).fill(1)
+    }
+    const keyring = Keyring.generate()
+    const roleBased = []
+    for (let i = 0; i < numArr.length; i++) {
+        const keys = []
+        for (let j = 0; j < numArr[i]; j++) {
+            keys.push(PrivateKey.generate())
+        }
+        roleBased.push(keys)
+    }
+    keyring.key = roleBased
+    return keyring
+}
+
 module.exports = {
     unitMap,
+    generateDecoupledKeyring,
+    generateMultiSigKeyring,
+    generateRoleBasedKeyring,
 }
